# api/openapi-gus.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "BDL API",
      "version": "v1"
    },
    "servers": [
      {
        "url": "/api/v1"
      }
    ],
    "paths": {
...
      "/subjects": {
        "get": {
          "tags": [
            "Subjects"
          ],
          "summary": "Lista tematów / Subjects list",
          "parameters": [
            {
              "name": "parent-id",
              "in": "query",
              "description": "Identyfikator tematu nadrzędnego / Parent subject's Id",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "page",
              "in": "query",
              "description": "Indeks strony / Page index",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            },
            {
              "name": "page-size",
              "in": "query",
              "description": "Rozmiar strony / Page size",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 10
              }
            },
            {
              "name": "sort",
              "in": "query",
              "description": "Oczekiwana kolejność / Desired order",
              "schema": {
                "enum": [
                  "Id",
                  "-Id",
                  "Id,Name",
                  "Id,-Name",
                  "-Id,Name",
                  "-Id,-Name",
                  "Name",
                  "-Name",
                  "Name,Id",
                  "Name,-Id",
                  "-Name,Id",
                  "-Name,-Id"
                ],
                "type": "string"
              }
            },
            {
              "name": "lang",
              "in": "query",
              "description": "Oczekiwany język odpowiedzi - opcjonalny (jeśli podano parametr, nagłówek \"Accept-Language\" jest ignorowany) / Expected response conent language - optional (if parameter specified, request header \"Accept-Language\" is ignored)",
              "schema": {
                "enum": [
                  "pl",
                  "en"
                ],
                "type": "string"
              }
            },
            {
              "name": "format",
              "in": "query",
              "description": "Oczekiwany format odpowiedzi - opcjonalny (jeśli podano parametr, nagłówek \"Accept\" jest ignorowany) / Expected response content type - optional (if parameter specified, request header \"Accept\" is ignored)",
              "schema": {
                "enum": [
                  "json",
                  "jsonapi",
                  "xml"
                ],
                "type": "string"
              }
            },
            {
              "name": "Accept-Language",
              "in": "header",
              "description": "Oczekiwany język odpowiedzi - opcjonalny (jeśli podano parametr \"lang\", nagłówek \"Accept-Language\" zostanie ignorowany) / Expected response conent language - optional (if \"lang\" parameter is specified, the \"Accept-Language\" header will be ignored)",
              "schema": {
                "enum": [
                  "pl",
                  "en"
                ],
                "type": "string"
              }
            },
            {
              "name": "Accept",
              "in": "header",
              "description": "Oczekiwany format odpowiedzi - opcjonalny (jeśli podano parametr \"format\", nagłówek \"Accept\" zostanie zignorowany) / Expected response content type - optional (if the \"format\" parameter is specified, the \"Accept\" header will be ignored)",
              "schema": {
                "enum": [
                  "application/json",
                  "application/vnd.api+json",
                  "application/xml"
                ],
                "type": "string"
              }
            },
            {
              "name": "If-None-Match",
              "in": "header",
              "description": "Nagłówek warunkowego żadania If-None-Match (entity tag)/Conditional Requests header If-None-Match (entity tag)",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "If-Modified-Since",
              "in": "header",
              "description": "Nagłówek warunkowego żadania If-Modified-Since/Conditional Requests header If-Modified-Since",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "500": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseMessage"
                  }
                },
                "application/vnd.api+json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseMessage"
                  }
                },
                "application/xml": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseMessage"
                  }
                }
              }
            },
            "404": {
              "description": "Not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseMessage"
                  }
                },
                "application/vnd.api+json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseMessage"
                  }
                },
                "application/xml": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseMessage"
                  }
                }
              }
            },
            "200": {
              "description": "Success",
              "headers": {
                "ETag": {
                  "description": "Entity tag",
                  "schema": {
                    "type": "string"
                  }
                },
                "Date": {
                  "description": "Server date and time",
                  "schema": {
                    "type": "string"
                  }
                },
                "X-Rate-Limit-Limit": {
                  "description": "Rate limit limit",
                  "schema": {
                    "type": "string"
                  }
                },
                "X-Rate-Limit-Remaining": {
                  "description": "Rate limit remaining",
                  "schema": {
                    "type": "string"
                  }
                },
                "X-Rate-Limit-Reset": {
                  "description": "Rate limit reset",
                  "schema": {
                    "type": "string"
                  }
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SubjectList"
                  }
                },
                "application/vnd.api+json": {
                  "schema": {
                    "$ref": "#/components/schemas/SubjectList"
                  }
                },
                "application/xml": {
                  "schema": {
                    "$ref": "#/components/schemas/SubjectList"
                  }
                }
              }
            },
            "429": {
              "$ref": "#/components/responses/LimitResponseMessage"
            }
          }
        }
      },
...
      "/subjects/{id}": {
        "get": {
          "tags": [
            "Subjects"
          ],
          "summary": "Temat o zadanym / Subject with selected Id",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Id tematu / Subject Id",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "lang",
              "in": "query",
              "description": "Oczekiwany język odpowiedzi - opcjonalny (jeśli podano parametr, nagłówek \"Accept-Language\" jest ignorowany) / Expected response conent language - optional (if parameter specified, request header \"Accept-Language\" is ignored)",
              "schema": {
                "enum": [
                  "pl",
                  "en"
                ],
                "type": "string"
              }
            },
            {
              "name": "format",
              "in": "query",
              "description": "Oczekiwany format odpowiedzi - opcjonalny (jeśli podano parametr, nagłówek \"Accept\" jest ignorowany) / Expected response content type - optional (if parameter specified, request header \"Accept\" is ignored)",
              "schema": {
                "enum": [
                  "json",
                  "jsonapi",
                  "xml"
                ],
                "type": "string"
              }
            },
            {
              "name": "Accept-Language",
              "in": "header",
              "description": "Oczekiwany język odpowiedzi - opcjonalny (jeśli podano parametr \"lang\", nagłówek \"Accept-Language\" zostanie ignorowany) / Expected response conent language - optional (if \"lang\" parameter is specified, the \"Accept-Language\" header will be ignored)",
              "schema": {
                "enum": [
                  "pl",
                  "en"
                ],
                "type": "string"
              }
            },
            {
              "name": "Accept",
              "in": "header",
              "description": "Oczekiwany format odpowiedzi - opcjonalny (jeśli podano parametr \"format\", nagłówek \"Accept\" zostanie zignorowany) / Expected response content type - optional (if the \"format\" parameter is specified, the \"Accept\" header will be ignored)",
              "schema": {
                "enum": [
                  "application/json",
                  "application/vnd.api+json",
                  "application/xml"
                ],
                "type": "string"
              }
            },
            {
              "name": "If-None-Match",
              "in": "header",
              "description": "Nagłówek warunkowego żadania If-None-Match (entity tag)/Conditional Requests header If-None-Match (entity tag)",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "If-Modified-Since",
              "in": "header",
              "description": "Nagłówek warunkowego żadania If-Modified-Since/Conditional Requests header If-Modified-Since",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "500": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseMessage"
                  }
                },
                "application/vnd.api+json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseMessage"
                  }
                },
                "application/xml": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseMessage"
                  }
                }
              }
            },
            "404": {
              "description": "Not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseMessage"
                  }
                },
                "application/vnd.api+json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseMessage"
                  }
                },
                "application/xml": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseMessage"
                  }
                }
              }
            },
            "200": {
              "description": "Success",
              "headers": {
                "ETag": {
                  "description": "Entity tag",
                  "schema": {
                    "type": "string"
                  }
                },
                "Date": {
                  "description": "Server date and time",
                  "schema": {
                    "type": "string"
                  }
                },
                "X-Rate-Limit-Limit": {
                  "description": "Rate limit limit",
                  "schema": {
                    "type": "string"
                  }
                },
                "X-Rate-Limit-Remaining": {
                  "description": "Rate limit remaining",
                  "schema": {
                    "type": "string"
                  }
                },
                "X-Rate-Limit-Reset": {
                  "description": "Rate limit reset",
                  "schema": {
                    "type": "string"
                  }
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SubjectDetails"
                  }
                },
                "application/vnd.api+json": {
                  "schema": {
                    "$ref": "#/components/schemas/SubjectDetails"
                  }
                },
                "application/xml": {
                  "schema": {
                    "$ref": "#/components/schemas/SubjectDetails"
                  }
                }
              }
            },
            "429": {
              "$ref": "#/components/responses/LimitResponseMessage"
            }
          }
        }
      },
...
    },
    "components": {
      "responses": {
        "LimitResponseMessage": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LimitResponseMessage"
              }
            }
          },
          "description": "Definicja błędu limitu / Limit error definition"
        }
      },
      "schemas": {
...
        "Available": {
          "type": "object",
          "properties": {
            "year": {
              "type": "string",
              "description": "Rok / Year",
              "nullable": true
            },
            "levels": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              },
              "description": "Poziomy dostępności / Availability levels",
              "nullable": true
            },
            "quarterly": {
              "type": "string",
              "description": "Dane roczne czy kwartalne / Yearly or quarterly data",
              "nullable": true
            }
          },
          "additionalProperties": false,
          "description": "Opis dostępności w roku / Availability in year description"
        },
        "ErrorMessage": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "description": "Opis błędu / Error result",
              "nullable": true
            },
            "reason": {
              "type": "string",
              "description": "Opis przyczyny błędu / Error reason description",
              "nullable": true
            },
            "solution": {
              "type": "string",
              "description": "Rozwiązanie problemu / Error solution",
              "nullable": true
            },
            "code": {
              "type": "integer",
              "description": "Kod błędu / Error code",
              "format": "int32",
              "nullable": true
            },
            "help": {
              "type": "string",
              "description": "Link do pomocy / Help link",
              "nullable": true
            }
          },
          "additionalProperties": false,
          "description": "Komunikat błędu / Error message"
        },
        "ErrorResponseMessage": {
          "type": "object",
          "properties": {
            "errors": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ErrorMessage"
              },
              "description": "Lista błedów / Erros list",
              "nullable": true
            }
          },
          "additionalProperties": false,
          "description": "Definicja listy błędów / Errors list definition"
        },
...
        "LimitResponseMessage": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "description": "Opis błędu / Error result",
              "nullable": true
            },
            "reason": {
              "type": "string",
              "description": "Opis przyczyny błędu / Error reason description",
              "nullable": true
            }
          },
          "additionalProperties": false,
          "description": "Definicja błędu limitu / Limit error definition"
        },
...
        "Metadata": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "Id metadanych / Matadata's id",
              "nullable": true
            },
            "title": {
              "type": "string",
              "description": "Nazwa zasobu / Resource name",
              "nullable": true
            },
            "url": {
              "type": "string",
              "description": "Url zasobu / Resource Url",
              "nullable": true
            },
            "provider": {
              "type": "string",
              "description": "Dostawca zasobu / Resource provider",
              "nullable": true
            },
            "modificationDate": {
              "type": "string",
              "description": "Data ostatniej modyfikacji / Last modification date",
              "format": "date-time"
            },
            "description": {
              "type": "string",
              "description": "Opis zasobu / Resource description",
              "nullable": true
            },
            "keywords": {
              "type": "string",
              "description": "Słowa kluczowe / Describing keywords",
              "nullable": true
            },
            "temporalCoverage": {
              "type": "string",
              "description": "",
              "nullable": true
            },
            "language": {
              "type": "string",
              "description": "Języki prezentowania zasobu / Resource's languages",
              "nullable": true
            },
            "updatePeriod": {
              "type": "string",
              "description": "Częstotliwość aktualizacji / Update frequency",
              "nullable": true
            },
            "contentType": {
              "type": "string",
              "description": "Typ treści / Content type",
              "nullable": true
            }
          },
          "additionalProperties": false,
          "description": "Definicja metadanych zasobu / Resource metadata definition"
        },
        "PageLinks": {
          "type": "object",
          "properties": {
            "firstPage": {
              "type": "string",
              "description": "Pierwsza strona / First page",
              "nullable": true
            },
            "prevPage": {
              "type": "string",
              "description": "Poprzednia strona / Previous page",
              "nullable": true
            },
            "self": {
              "type": "string",
              "description": "Bieżąca strona / Current page",
              "nullable": true
            },
            "nextPage": {
              "type": "string",
              "description": "Następna strona / Next page",
              "nullable": true
            },
            "lastPage": {
              "type": "string",
              "description": "Ostatnia strona / Last page",
              "nullable": true
            }
          },
          "additionalProperties": false,
          "description": "Linki do stron wyników / Result page links"
        },
...
        "Subject": {
          "type": "object",
          "properties": {
            "metadata": {
              "$ref": "#/components/schemas/Metadata"
            },
            "version": {
              "$ref": "#/components/schemas/Version"
            },
            "id": {
              "type": "string",
              "description": "Identyfikator tematu / Subject identifier",
              "nullable": true
            },
            "parentId": {
              "type": "string",
              "description": "Identyfikator tematu nadrzędnego / Parent subject identifier",
              "nullable": true
            },
            "name": {
              "type": "string",
              "description": "Nazwa tematu / Subject name",
              "nullable": true
            },
            "hasVariables": {
              "type": "boolean",
              "description": "Czy temat posiada zmienne / Does the subject has variables"
            },
            "children": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Identyfikatory tematów podrzędnych / Children subject identifiers",
              "nullable": true
            },
            "levels": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              },
              "description": "Poziomy agregacji tematu / Subject agregate levels",
              "nullable": true
            }
          },
          "additionalProperties": false,
          "description": "Reprezentacja tematu / Subjects representation"
        },
        "SubjectDetails": {
          "type": "object",
          "properties": {
            "metadata": {
              "$ref": "#/components/schemas/Metadata"
            },
            "version": {
              "$ref": "#/components/schemas/Version"
            },
            "id": {
              "type": "string",
              "description": "Identyfikator tematu / Subject identifier",
              "nullable": true
            },
            "parentId": {
              "type": "string",
              "description": "Identyfikator tematu nadrzędnego / Parent subject identifier",
              "nullable": true
            },
            "name": {
              "type": "string",
              "description": "Nazwa tematu / Subject name",
              "nullable": true
            },
            "hasVariables": {
              "type": "boolean",
              "description": "Czy temat posiada zmienne / Does the subject has variables"
            },
            "children": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Identyfikatory tematów podrzędnych / Children subject identifiers",
              "nullable": true
            },
            "levels": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              },
              "description": "Poziomy agregacji tematu / Subject agregate levels",
              "nullable": true
            },
            "years": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              },
              "description": "Lata obowiązywania tematu / Years of subject validity",
              "nullable": true
            },
            "availability": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Available"
              },
              "description": "Opis dostępności w latach / Description of availability in years",
              "nullable": true
            },
            "dimensions": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Wymiary tematu / Subject dimensions",
              "nullable": true
            },
            "lastUpdate": {
              "type": "string",
              "description": "Data ostatniej aktualizacji tematu / Date of the last update of the subject",
              "format": "date-time",
              "nullable": true
            },
            "description": {
              "type": "string",
              "description": "Opis tematu / Subject description",
              "nullable": true
            }
          },
          "additionalProperties": false,
          "description": "Szczegółowa reprezentacja tematu / Detailed representation of the subject"
        },
        "SubjectList": {
          "type": "object",
          "properties": {
            "totalRecords": {
              "type": "integer",
              "description": "",
              "format": "int32",
              "nullable": true
            },
            "page": {
              "type": "integer",
              "description": "",
              "format": "int32",
              "nullable": true
            },
            "pageSize": {
              "type": "integer",
              "description": "",
              "format": "int32",
              "nullable": true
            },
            "links": {
              "$ref": "#/components/schemas/PageLinks"
            },
            "results": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Subject"
              },
              "description": "",
              "nullable": true
            },
            "metadata": {
              "$ref": "#/components/schemas/Metadata"
            },
            "version": {
              "$ref": "#/components/schemas/Version"
            }
          },
          "additionalProperties": false,
          "description": "Lista tematów / List of the subjects"
        },
...
        "Version": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "Pełny numer wersji / Full version number",
              "nullable": true
            },
            "major": {
              "type": "integer",
              "description": "Numer główny wersji / The main version number",
              "format": "int32"
            },
            "minor": {
              "type": "integer",
              "description": "Numer dodatkowy wersji / The additional version number",
              "format": "int32"
            },
            "modificationDate": {
              "type": "string",
              "description": "Data ostatniej modyfikacji / Date of the last modification",
              "format": "date-time"
            },
            "deprecated": {
              "type": "string",
              "description": "Data wycofania / Date of the deprecation",
              "format": "date-time",
              "nullable": true
            }
          },
          "additionalProperties": false,
          "description": "Reprezentacja wersji API BDL / Representation of the API BDL version"
        },
...
      }
    }
  }